requirements.txt
diffusers==0.21.1
numpy==1.24.4
spacy==3.6.1
tqdm==4.65.0
transformers==4.30.2
torch==2.0.1
accelerate==0.23.0
pandas==2.2.2
matplotlib==3.8.4
nltk==3.8.1
torchvision
torchaudio
facenet-pytorch
pillow
piq
opencv-python
easydict
scipy
insightface
boto3
onnxruntime-gpu
seaborn
wandb
scikit-image
kornia
GPUtil
tabulate
openai-clip==1.0.1
datasets
face-recognition